Fix type mismatch
VM::alloc() usuall follows memSize() which returns size_t.
It is only logical for alloc() to take a parameter of the same type, as
there is good reason - IMO - for it to take just uint32_t.

+++ vm/runtime.cpp
@@ -102,7 +102,7 @@ VM::VM()
 Allocates a block of memory
 Note that this function guarantees that the memory is zeroed out
 */
-Value VM::alloc(uint32_t size, Tag tag)
+Value VM::alloc(size_t size, Tag tag)
 {
     // FIXME: use an alloc pool of some kind
     auto ptr = static_cast<refptr>(calloc(1, size));
